Holy moly big as commit incoming
feat: add class variable to placements controller to hold the stages and layers and also get the stages and layers from teh drawer.js

feat: add a centre tile attribute to ships
refactor: modify setup input listeners to use new functions to place and return ships

Mega green incoming 👽👽👽👽👽

feat: add new update snape to tile function that takes a konvaship object as input and draws the ship accordingly to where the ship should snap into place on the grid when being dragged

feat: add new get ship cells from centre tile that is given a ship, and the centre x and y position. this processes the center tile to return an array of cells / tiles that the ship would occupy

feat: add function to validate ship placement and return true or false. checks if the ship is out of bounds or is overlapping another ship

feat: add place ship function that is given a ship and places it at the position of where the mouse pointer is currently hovering over